// src/app/enums/platform.enum.ts
export enum PlatformEnum {
  TWITCH = 'TWITCH',
  KICK = 'KICK',
  YOUTUBE = 'YOUTUBE',
  INSTAGRAM = 'INSTAGRAM',
  ALL = 'ALL'
}

export enum PlatformColorEnum {
  TWITCH = '#7734eb',
  KICK = '#34eb5c',
  YOUTUBE = '#e33939',
  ALL = '#3974e3',
  TWITCH_ALFA = 'rgba(119,52,235,0.47)',
  KICK_ALFA = 'rgba(52,235,92,0.4)',
  YOUTUBE_ALFA = 'rgba(227,57,57,0.43)',
  ALL_ALFA = 'rgba(57,116,227,0.42)',
  TWITCH_MAX_ALFA = 'rgba(119,85,177,0.45)',
  KICK_MAX_ALFA = 'rgba(52,235,92,0.16)',
  YOUTUBE_MAX_ALFA = 'rgba(227,57,57,0.2)',
}